Add CustomerAppealType to customers table and Nova
This commit introduces a new feature for customer profiles. A new enum 'appeal_type' was included in the database migration for the customers table. Moreover, the CustomerAppealType has been added to the Nova options for both 'renter' and 'passenger' customer types in the Customer Nova resource. This allows admin users to track and filter customers based on their preferred form of address.