"Add Route model and update Nova resources
The Route model is now integrated into the system, with the 'routes' table defined and its 'id' guarded. The Route is also added into the system via NovaServiceProvider as part of 'Kelionės' group. Moreover, 'distance' column was removed from 'routes' table as it was deemed unnecessary for current implementation."