Add Driver and User relationships to Trip model
Added new relationships to the Trip model to associate Drivers and Users to trips. RoleType enum was imported to only allow users with a driver role to be associated as a driver in a trip. Trips in Nova now include fields for assigning drivers and auxiliary users. The Trip table migration was also updated to include foreign key columns for driver_id and user_id for data integrity.