Add trip time format and data casting
Added a `formatTime` function and data type casting in the Trip model to better structure the Trip's time data. This enables the backend to store and process the trip's departure and arrival times in a uniform manner. Also, updated the Nova Trip resource to include these changes, thus resulting in a more intuitive and user-friendly interface.