Add payment methods and localize data seeding
The 'Trip' model has been expanded by adding 'arrived_at' and 'departure_at' to the 'dates' attribute and adding a new relationship with the 'PaymentMethod' model. A Nova field for the 'paymentMethod' attribute has been added into the Trip resource in Nova. The database seeding of bus types and features was also localized to enable internationalization. A new seeder for payment methods was introduced and called in the 'DatabaseSeeder'.