Add relation to Trip and TripCustomer models
A new relation has been added to both the Trip and TripCustomer models. Now, the `Trip` model contains a `participants` method that indicates a 'belongsToMany' relationship with the `Customer` class. This is reflected in the app Interface, which now provides a field for adding participants to a trip. In `TripCustomer`, the `$model` is also changed to `Customer` and `$title` is assigned to 'full_name'. These changes improve the usability of the app by allowing trip participants to be selected directly in the Interface.