Add call action to TripCustomer in Nova
In TripCustomer Nova Resource, a new action called 'PaymentTicketNeedCallAction' has been added to give the administrators an option to mark a customer for a call. A check is also implemented to show this action only when it's necessary. The commit addresses the need for a manual follow-up system in case of any issues with payment.